Set TERM and Locale vars properly
This commit sets the term and locale vars.  The locale is set before
isntalling anything related to postgres so that we are always using the
correct character encoding for anything postgres related.  We also purge
the previous locale and use the update-locale command to make sure the
change is reflected everywhere it is needed