Nov Demo - Full build
This branch has the Aether Nov Demo with all the required containers:  Aether + Logstash + Elasticsearch + Kibana.  It does not have Kafka.

To spin it up, use:
docker-compose -f docker-compose.yml -f docker-compose.demo.yml up --build

It's best to not use nuke but to spin up the containers and wait for them to stabalize.  It's pretty slow on a laptop vs AWS.

Use reset.sh from a different terminal to set things up.