AET-79: add /schema endpoint (#30)
Integrate drf-openapi for auto-generated documentation.

kernel: change versioning scheme from NamespaceVersioning to
URLPathVersioning. This is necessary since drf-openapi only
supports the latter. See
https://github.com/limdauto/drf_openapi#4-constraints

NOTE: this commit pins the django-restframework version to 3.6.4,
which is the highest available version which drf-openapi supports.
As soon as limdauto/drf_openapi#94 is
addressed, we can remove this version constraint.

Resolves https://jira.ehealthafrica.org/browse/AET-79