Limit the `count` query to 1
We don't need to know how many new items we have, we just need to know
whether we have new items. Besides, we were unwrapping them and
iterating over them in memory.

The parameter `topic` was removed because it was not being used. And the
ordering of elements was also removed because we don't need elements in
any particular order when we are just counting them, that only slows
down the query.