Add timeout parameter to client
It's important that no calls to any service leave the system
unresponsive, consuming resources for a long period. Timeouts by default
are counted in minutes, which is ridiculous for any web app.

If any call took longer a few seconds then it must be looked into, with
long timeouts we won't even know that there is a problem.