Remove seemingly irrelevant code
This piece of code is not testing anything, i.e. it's not asserting
either that expected parameters are accepted by the function, nor that
unrecognized parameters are not accepted. And even that test would be
irrelevant because that's not a functionality provided by the client,
but by the language itself.