Delete *.whl files
These are autogenerated binary files that should not be committed in the
repository. They get recreated at build time, no need to have them here.

Wheel files are the #1 source of merge/rebase conflicts, and since they
are binary it's impossible to know which one is the most up-to-date (if
any, most likely they will both contain necessary changes, so they have
to be recreated once again after the conflict is resolved).

There is only 1 module left as whl files, and it's the
`django_ums_client`. This is because this module is in an external
private repository, and getting pip to fetch it from travis is not
trivial. However, because it's in a different repository, this file
doesn't suffer from the same problems as all the others, so it's fine to
keep it here.