feat: avro validation with error accumulation (#217)
Improve validation of avro schemas -- accumulate a list of errors
encountered during validation instead of just returning a boolean.

Background: when attempting to validate a complex form in the
aether ui, any mapping which produces an entity with type errors
will result in a single generic error. This becomes a problem when
a form has more than one field -- the user does not get any hints as
to what went wrong (what type was expected? what was the actual
value? which part of the schema did not validate?).

The implementation of AvroValidator is based on the official
avro.io.validate() function, which can be found here:
https://github.com/apache/avro/blob/b71dcf24f252da5858bd95bfd4bd56402e3a458c/lang/py3/avro/io.py#L96-L142.

avro.io.validate() returns a boolean which
indicates the outcome of the validation. No error details
are returned. In contrast, AvroValidator will accumulate a
list of all errors encountered during validation. The type
of AvroValidator.errors is a list of AvroValidationErrors.

Performance: running AvroValidator.validate() will be about
twice as slow as avro.io.validate().
avro.io.validate() will check 10000 datums against a
moderately complex schema in ~0.5 seconds. AvroValidator
will finish in ~1 second. If this is not acceptable, we could
add a full_validation= flag in kernel, the value of
which would only be true when called from the ui module.