Add timeout parameter to client (#221)
* Add timeout parameter to client

It's important that no calls to any service leave the system
unresponsive, consuming resources for a long period. Timeouts by default
are counted in minutes, which is ridiculous for any web app.

If any call took longer a few seconds then it must be looked into, with
long timeouts we won't even know that there is a problem.

* Remove seemingly irrelevant code

This piece of code is not testing anything, i.e. it's not asserting
either that expected parameters are accepted by the function, nor that
unrecognized parameters are not accepted. And even that test would be
irrelevant because that's not a functionality provided by the client,
but by the language itself.

* Rename `client` to `_client` in tests

Each test shadows the original client with its own local variable
`client`.