fix: remote storage, file urls (#335)
* Fix file urls in kernel and odk.

* Add function resolve_file_url() to common module.

For local development, the environment variable
DJANGO_STORAGE_BACKEND is set to "filesystem" and files
(attachments, media files etc.) are stored on the file
system and served via nginx. Example URL:
http://odk.aether.local/media/<path-to-file>. The hostname
of the docker container is prepended to the media file url
so that the resource can be served locally.

When the environment variable DJANGO_STORAGE_BACKEND is set
to "s3" or "gcs", all file urls will be absolute. Example:
https://abcd.s3.amazonaws.com/<file-name>?AWSAccessKeyId=ABC&Signature=ABC%3D&Expires=1534775613.
This url is not modified.

* Adjust ODK manifest file creation so that it works with
  remote storage.