feat: filter submissions by payload content
Add a dynamic filter to SubmissionViewSet. Any query
parameter prefixed by "payload__" will filter submissions
based on the contents of Submission.payload.

Example:
The URL "/submissions?payload__a__b__c=1" will yield
the queryset

  models.Submission.objects.filter(payload__a__b__c=1)

and return a list of submissions with a JSON payload like

  '{"a": {"b": {"c": 1}}}'

Note that it is possible to compare not just strings, but
numbers, lists and objects as well.