feat: filter submissions by payload content (#412)
Add a dynamic filter to SubmissionViewSet and EntityViewSet.
Any query parameter prefixed by "payload__" will filter 
submissions based on the contents of e.g. 
`models.Submission.payload`.

The reason for adding this to `views.py` (instead of `filters.py`,
where one would expect to find it) is that I wasn't able to tweak 
django filters to accept dynamic field names. Dynamic field
names are necessary, since we don't know the shape of the 
JSON payload in advance. 

Example:
The URL "/submissions?payload__a__b__c=1" will yield
the queryset

  models.Submission.objects.filter(payload__a__b__c=1)

and return a list of submissions with a JSON payload like

  '{"a": {"b": {"c": 1}}}'

Note that it is possible to compare not just strings, but
numbers, lists and objects as well.